FAC-397:feature:Refactor AssessmentUserRole
