+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/common/AssessmentUserRoleTest.java
@@ -1,13 +1,14 @@
 package org.flickit.assessment.core.common;
 
-import org.flickit.assessment.common.permission.AssessmentPermission;
+import org.flickit.assessment.common.application.assessment.AssessmentPermission;
+import org.flickit.assessment.core.application.domain.AssessmentUserRole;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 import java.util.Set;
 
-import static org.flickit.assessment.common.permission.AssessmentPermission.VIEW_ASSESSMENT;
-import static org.flickit.assessment.core.common.AssessmentUserRole.*;
+import static org.flickit.assessment.common.application.assessment.AssessmentPermission.VIEW_ASSESSMENT;
+import static org.flickit.assessment.core.application.domain.AssessmentUserRole.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 class AssessmentUserRoleTest {
